add fobos and lr, softmax reg

+++ oml/functions.py
@@ -43,7 +43,7 @@ def cross_entropy(y, t):
         t = t.argmax(axis=1)
 
     batch_size = y.shape[0]
-    return -np.sum(np.log(y[np.arange(batch_size), t])) / batch_size
+    return -np.sum(np.log(y[np.arange(batch_size), t.astype(int)])) / batch_size
 
 
 class StrongConvexity:
